requirements.txt
numpy==1.11.0
scipy==1.1.0
torchvision==0.2.1
opencv_python==3.3.0.10
easydict==1.7
torch==0.4.1
Pillow==6.0.0
motmetrics==1.1.3
PyYAML==5.1
